structure/requirements.txt
python-dotenv==1.0.*
griptape
requests
psycopg2-binary